requirements.txt
opencv-python>=4.8.0
openai-whisper>=20230314
pillow>=10.0.0
numpy>=1.24.0
ffmpeg-python>=0.2.0
anthropic>=0.7.0
yt-dlp>=2023.10.13
streamlit>=1.28.0
requests>=2.31.0
ollama>=0.1.0
google-generativeai>=0.3.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0
